may 12

+++ data/createsummer.py
@@ -44,7 +44,7 @@ def printtable():
         for line in dfile.read().split('\n'):
             a = line.split(',')
             kscore = round(max(0, getkattisscore(a[1]) - float(a[3])), 2)
-            cscore = float(a[4])
+            cscore = int(a[4])
             data.append((a[0], kscore, cscore, kscore+cscore))
     data.sort(key=lambda x: -1*x[3])
     for a in data:
